requirements.txt
PyYAML
openai